media-fonts/oswald: more precise version for live ebuild

+++ media-fonts/oswald/oswald-3.0.9999.ebuild
@@ -32,6 +32,8 @@ IUSE=""
 RDEPEND="${DEPEND}"
 
 src_prepare() {
-	find "${S}" -mindepth 2 -type f -name 'Oswald-*.[ot]tf' -exec mv {} "${S}" \;
-	rm -f *unhinted.ttf
+	if [[ ${PV} == *9999* ]]; then
+		find "${S}"/${PV%.*} -type f -name '*.[ot]tf' -exec mv {} "${S}" \;
+		rm -f *unhinted.ttf
+	fi
 }
